Disabled group3 interrupts when in the i2c code to prevent re-entrancy
A potential issue for a re-entracy lock existed if the i2c code was
being executed when a interrupt fired. when i2c_rx or i2c_tx encountered
the while loop that checks for the bus being free it would have locked.

Prevented the issue by disabling group 3 interrupts while making an i2c
tx or rx.